+++ src/hasClass.js
@@ -1,0 +1,11 @@
+'use strict'
+
+/**
+ * [hasClass description]
+ * @param  {[type]}  el            [description]
+ * @param  {[type]}  ...classNames [description]
+ * @return {Boolean}               [description]
+ */
+export default function hasClass (el, ...classNames) {
+  return el.classList.contains(...classNames)
+}
