+++ src/off.js
@@ -1,0 +1,12 @@
+'use strict'
+
+/**
+ * [off description]
+ * @param  {[type]} el        [description]
+ * @param  {[type]} eventName [description]
+ * @param  {[type]} func      [description]
+ * @return {[type]}           [description]
+ */
+export default function off (el, eventName, func) {
+  el.removeEventListener(eventName, func)
+}
